Move grammaermenu components to seperate folder and update snapshots
